revert server context handling
Former-commit-id: 97b1fb82742c15f8796e04c8c8d0589b09b17abd

+++ modules/Platform/edm_qm/WindowsFunctionsForLinux.cpp
@@ -2,7 +2,7 @@
 #include <sys/stat.h>
 #include "WindowsFunctionsForLinux.h"
 
-unsigned int GetTickCount()
+unsigned long GetTickCount()
 {
    struct timeval tv;
    if (gettimeofday(&tv, 0) != 0)
